Add not found exception for LinkHubEditor

+++ backend/shrunk/api/linkhub.py
@@ -49,10 +49,12 @@ def create_linkhub(netid: str, client: ShrunkClient, req: Any) -> Any:
 
 @bp.route("/<string:linkhub_id>", methods=["GET"])
 @require_login
-def get_linkhub_by_id(_netid: str, client: ShrunkClient, linkhub_id: str) -> Any:
+def get_linkhub_by_id(netid: str, client: ShrunkClient, linkhub_id: str) -> Any:
     result = client.linkhubs.get_by_id(linkhub_id)
     if result is None:
-        return jsonify({"error": "does not exist"})  # TODO: make cleaner.
+        return jsonify(
+            {"success": False, "error": "Does not exist"}, 404
+        )  # TODO: make cleaner.
 
     return jsonify(result)
 
@@ -72,9 +74,11 @@ ADD_LINK_TO_LINKHUB_SCHEMA = {
 @request_schema(ADD_LINK_TO_LINKHUB_SCHEMA)
 @require_login
 def add_link_to_linkhub(
-    _netid: str, client: ShrunkClient, req: Any, linkhub_id: str
+    netid: str, client: ShrunkClient, req: Any, linkhub_id: str
 ) -> Any:
-    # TODO: Add security.
+    if not client.linkhubs.can_edit(linkhub_id, netid):
+        return jsonify({"success": False, "error": "No permission"}), 401
+
     client.linkhubs.add_link_to_linkhub(linkhub_id, req["title"], req["url"])
 
     return jsonify({"success": True})
@@ -96,9 +100,10 @@ SET_CHANGE_LINKHUB_TITLE_SCHEMA = {
 @request_schema(SET_CHANGE_LINKHUB_TITLE_SCHEMA)
 @require_login
 def set_link_from_linkhub(
-    _netid: str, client: ShrunkClient, req: Any, linkhub_id: str
+    netid: str, client: ShrunkClient, req: Any, linkhub_id: str
 ) -> Any:
-    # TODO: Add security.
+    if not client.linkhubs.can_edit(linkhub_id, netid):
+        return jsonify({"success": False, "error": "No permission"}), 401
 
     client.linkhubs.set_data_at_link_to_linkhub(
         linkhub_id, req["index"], req["title"], req["url"]
@@ -121,9 +126,10 @@ CHANGE_LINKHUB_TITLE_SCHEMA = {
 @request_schema(CHANGE_LINKHUB_TITLE_SCHEMA)
 @require_login
 def change_linkhub_title(
-    _netid: str, client: ShrunkClient, req: Any, linkhub_id: str
+    netid: str, client: ShrunkClient, req: Any, linkhub_id: str
 ) -> Any:
-    # TODO: Add security.
+    if not client.linkhubs.can_edit(linkhub_id, netid):
+        return jsonify({"success": False, "error": "No permission"}), 401
 
     client.linkhubs.change_title(linkhub_id, req["title"])
 
@@ -144,9 +150,10 @@ CHANGE_LINKHUB_TITLE_SCHEMA = {
 @request_schema(CHANGE_LINKHUB_TITLE_SCHEMA)
 @require_login
 def change_linkhub_alias(
-    _netid: str, client: ShrunkClient, req: Any, linkhub_id: str
+    netid: str, client: ShrunkClient, req: Any, linkhub_id: str
 ) -> Any:
-    # TODO: Add security.
+    if not client.linkhubs.can_edit(linkhub_id, netid):
+        return jsonify({"success": False, "error": "No permission"}), 401
 
     client.linkhubs.change_alias(linkhub_id, req["alias"])
 
@@ -167,9 +174,10 @@ DELETE_LINK_FROMLINKHUB_SCHEMA = {
 @request_schema(DELETE_LINK_FROMLINKHUB_SCHEMA)
 @require_login
 def delete_link_from_linkhub(
-    _netid: str, client: ShrunkClient, req: Any, linkhub_id: str
+    netid: str, client: ShrunkClient, req: Any, linkhub_id: str
 ) -> Any:
-    # TODO: Add security.
+    if not client.linkhubs.can_edit(linkhub_id, netid):
+        return jsonify({"success": False, "error": "No permission"}), 401
 
     client.linkhubs.delete_element_at_index(linkhub_id, req["index"])
